Move repos to folder that matches namespace
