+++ src/xluhco.web/Repositories/IShortLinkRepository.cs
@@ -1,6 +1,6 @@
 ﻿using System.Collections.Generic;
 
-namespace xluhco.web
+namespace xluhco.web.Repositories
 {
     public interface IShortLinkRepository
     {
