+++ src/xluhco.web/Startup.cs
@@ -4,6 +4,7 @@ using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Logging;
 using Serilog;
+using xluhco.web.Repositories;
 
 namespace xluhco.web
 {
